use common run_async_function
